refactor(core): reorg connector files
