+++ packages/core/src/libraries/sign-in-experience/sign-in.ts
@@ -1,9 +1,9 @@
 import type { SignIn, SignUp } from '@logto/schemas';
 import { ConnectorType, SignInIdentifier } from '@logto/schemas';
 
-import type { LogtoConnector } from '#src/connectors/types.js';
 import RequestError from '#src/errors/RequestError/index.js';
 import assertThat from '#src/utils/assert-that.js';
+import type { LogtoConnector } from '#src/utils/connectors/types.js';
 
 export const validateSignIn = (
   signIn: SignIn,
