+++ packages/core/src/libraries/sign-in-experience/sign-up.ts
@@ -1,9 +1,9 @@
 import type { SignUp } from '@logto/schemas';
 import { SignInIdentifier, ConnectorType } from '@logto/schemas';
 
-import type { LogtoConnector } from '#src/connectors/types.js';
 import RequestError from '#src/errors/RequestError/index.js';
 import assertThat from '#src/utils/assert-that.js';
+import type { LogtoConnector } from '#src/utils/connectors/types.js';
 
 export const validateSignUp = (signUp: SignUp, enabledConnectors: LogtoConnector[]) => {
   for (const identifier of signUp.identifiers) {
