+++ packages/core/src/routes/interaction/utils/find-user-by-identifier.ts
@@ -1,4 +1,4 @@
-import { getLogtoConnectorById } from '#src/connectors/index.js';
+import { getLogtoConnectorById } from '#src/libraries/connector.js';
 import {
   findUserByEmail,
   findUserByUsername,
