+++ packages/core/src/routes/interaction/verifications/profile-verification-protected-identifier.test.ts
@@ -15,7 +15,7 @@ await mockEsmWithActual('#src/queries/user.js', () => ({
   hasUserWithIdentity: jest.fn().mockResolvedValue(false),
 }));
 
-mockEsm('#src/connectors/index.js', () => ({
+mockEsm('#src/libraries/connector.js', () => ({
   getLogtoConnectorById: jest.fn().mockResolvedValue({
     metadata: { target: 'logto' },
   }),
