+++ packages/core/src/routes/session/social.test.ts
@@ -4,8 +4,8 @@ import type { User } from '@logto/schemas';
 import Provider from 'oidc-provider';
 
 import { mockLogtoConnectorList, mockSignInExperience, mockUser } from '#src/__mocks__/index.js';
-import { getLogtoConnectorById } from '#src/connectors/index.js';
 import RequestError from '#src/errors/RequestError/index.js';
+import { getLogtoConnectorById } from '#src/libraries/connector.js';
 import { createRequester } from '#src/utils/test-utils.js';
 
 import socialRoutes, { signInRoute } from './social.js';
@@ -90,7 +90,7 @@ const getLogtoConnectorByIdHelper = jest.fn(async (connectorId: string) => {
   };
 });
 
-jest.mock('#src/connectors.js', () => ({
+jest.mock('#src/libraries/connector.js', () => ({
   getLogtoConnectors: jest.fn(async () => mockLogtoConnectorList),
   getLogtoConnectorById: jest.fn(async (connectorId: string) => {
     const connector = await getLogtoConnectorByIdHelper(connectorId);
