+++ packages/core/src/utils/connectors/index.ts
@@ -5,19 +5,6 @@ import path from 'path';
 import type { AllConnector, BaseConnector } from '@logto/connector-kit';
 import { ConnectorError, ConnectorErrorCodes, ConnectorType } from '@logto/connector-kit';
 
-import RequestError from '#src/errors/RequestError/index.js';
-import { findAllConnectors } from '#src/queries/connector.js';
-import assertThat from '#src/utils/assert-that.js';
-
-export const getConnectorConfig = async (id: string): Promise<unknown> => {
-  const connectors = await findAllConnectors();
-  const connector = connectors.find((connector) => connector.id === id);
-
-  assertThat(connector, new RequestError({ code: 'entity.not_found', id, status: 404 }));
-
-  return connector.config;
-};
-
 export function validateConnectorModule(
   connector: Partial<BaseConnector<ConnectorType>>
 ): asserts connector is BaseConnector<ConnectorType> {
